+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-None.
+### Fixed
+
+- Only interrupt due to errors when credo's exit status is ENOENT, 126, or 127, or when credo logs something to stderr
 
 ## [0.4.2] - 2021-04-16